restructure badge rendering & update badge creation to visualize all languages
