+++ source-code/badge/src/router.ts
@@ -2,37 +2,6 @@ import express, { Request, Response } from "express"
 import { Resvg } from "@resvg/resvg-js"
 import { badge } from "./badge.js"
 
-interface TranslationData {
-	[language: string]: number
-}
-
-interface ChartData {
-	labels: string[]
-	datasets: {
-		label: string
-		data: number[]
-		backgroundColor: string
-	}[]
-}
-
-// Function to generate the chart using Chart.js
-const generateChart = (data: TranslationData): string => {
-	const chartData: ChartData = {
-		labels: Object.keys(data),
-		datasets: [
-			{
-				label: "Translation Progress",
-				data: Object.values(data),
-				backgroundColor: "rgba(16, 185, 129, 0.7)",
-			},
-		],
-	}
-
-	// Chart.js code here...
-	// Return the chart as a base64 encoded string
-	return "base64-encoded-chart-string"
-}
-
 // Create the express router
 export const router = express.Router()
