+++ source-code/badge/src/types.d.ts
@@ -1,0 +1,5 @@
+export interface Percentage {
+	lang: string
+	percentage: number
+	count: { total: number; lint: number }
+}
